requirements.txt
# Indian Sign Language (ISL) Recognition
# Hugging Face Spaces Deployment
# Python 3.10+ required

# ========================================
# Core ML/DL Framework
# ========================================
tensorflow-cpu>=2.12.0

# ========================================
# Data Processing
# ========================================
numpy>=1.23.0
pillow>=9.5.0

# ========================================
# Gradio (for HF Spaces)
# ========================================
gradio>=4.0.0
